\name{ggradarSIZE-package}
\alias{ggradarSIZE-package}
\alias{ggradarSIZE}
\docType{package}
\title{Draw Color Radar Charts}
\description{Draw Colored Radar Charts using numeric data frames. This package was derived from the ggradar package, written by Ricardo Bion. Some tiny modifications were included to the original package in order to color the radar lines according to a weigth (or size) feature (required) that is supposed to be a numeric continuous variable}
\author{Damiano Fantini}
